+++ CHANGELOG.md
@@ -7,6 +7,10 @@ We go to the next version after each release on the
 
 ## [Unreleased]
 
+### Added
+
+- Backlight Toggle Button Visibility in Quick Settings.
+
 ### Fixed
 
 - On Screen Display (OSD) visibility on GNOME Shell 49.